+++ packages/tsurlfilter/cli/index.ts
@@ -6,6 +6,7 @@ import path from 'node:path';
 import { convertFilters, LOCAL_METADATA_FILE_NAME } from './convertFilters';
 import { version } from '../package.json';
 import { Extractor } from './extractFilters';
+import { generateMD5Hash } from '../src/utils/checksum';
 
 export const DEFAULT_DEST_RULE_SETS_DIR = './build/rulesets';
 
@@ -88,4 +89,4 @@ if (isRunningViaCli) {
 }
 
 // For API-like usage, we export the convertFilters function.
-export { convertFilters };
+export { convertFilters, generateMD5Hash };