+++ packages/tsurlfilter/src/rules/declarative-converter-utils/index.ts
@@ -4,3 +4,7 @@ export {
     getRuleSetPath,
     extractRuleSetId,
 } from './rule-set-path';
+
+export {
+    isSafeRule,
+} from './is-safe-dnr-rule';